Traverser now uses Java shorts to represent the number of loops. No one will EVER (EVER) do 32000+ loops.

+++ gremlin-core/src/main/java/com/tinkerpop/gremlin/process/Traverser.java
@@ -32,7 +32,7 @@ public interface Traverser<T> extends Serializable {
      *
      * @return The number of times the traverser has gone through a loop
      */
-    public int getLoops();
+    public short getLoops();
 
     /**
      * A traverser may represent a grouping of traversers to allow for more efficient data propagation.
